resolved-symbols: discover called functions
This application goes through the ELF libraries and binaries that a
process has mapped to memory, and for each such ELF file, it discovers
whether the undefined symbols (those from the PLT table) have yet been
resolved.

Symbol resolution is by default lazy, ie. the actual address is resolved
when the external function is called for the first time. In other words,
with this tool it is possible to determine what external functions have
been called at least once from each ELF library and binary.

Signed-off-by: Tommi Rantala <tt.rantala@gmail.com>